+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # svelte-morphing-modal changelog
 
+## 1.3.0
+* Implement scroll disabling when modal is open & `lockScroll` prop.
+* Rename `change` event to `toggle` event.
+* Dispatch `adjust` event when modal has switched between fullscreen and window mode
+* Blur effect for overlay.
+
 ## 1.2.1
 * Replace custom `blur` transition with new built-in. 
 * Dispatch `change` event any time open state changed.